+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removd
+
+- Remove `orgpermissions` controller as it is deprecated and all its resources have been cleaned up.
+
 ## [0.24.0] - 2022-03-30
 
 ### Added